pydantic_ai_slim/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pydantic-ai-slim"
version = "0.0.20"
description = "Agent Framework / shim to use Pydantic with LLMs, slim package"
authors = [
    { name = "Samuel Colvin", email = "samuel@pydantic.dev" },
]
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Unix",
    "Operating System :: POSIX :: Linux",
    "Environment :: Console",
    "Environment :: MacOS X",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet",
]
requires-python = ">=3.9"
dependencies = [
    "eval-type-backport>=0.2.0",
    "griffe>=1.3.2",
    "httpx>=0.27",
    "logfire-api>=1.2.0",
    "pydantic>=2.10",
]

[project.optional-dependencies]
# WARNING if you add optional groups, please update docs/install.md
logfire = ["logfire>=2.3"]
graph = ["pydantic-graph==0.0.20"]
openai = ["openai>=1.59.0"]
cohere = ["cohere>=5.13.11"]
vertexai = ["google-auth>=2.36.0", "requests>=2.32.3"]
anthropic = ["anthropic>=0.40.0"]
groq = ["groq>=0.12.0"]
mistral = ["mistralai>=1.2.5"]
mlx-lm = ["mlx-lm>=0.21.1"]

[dependency-groups]
dev = [
    "anyio>=4.5.0",
    "devtools>=0.12.2",
    "coverage[toml]>=7.6.2",
    "dirty-equals>=0.9.0",
    "inline-snapshot>=0.19.3",
    "pytest>=8.3.3",
    "pytest-examples>=0.0.14",
    "pytest-mock>=3.14.0",
    "pytest-pretty>=1.2.0",
    "diff-cover>=9.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["pydantic_ai"]

[tool.uv.sources]
pydantic-graph = { workspace = true }
